Initial implementation of hopscotch hashing algorithm.
The hashing alrorighthm has been changed to use open addressing in order
to avoid memory overhead of chaining algorithms.

The collision resolution method has been chosen to be hopscotch
algorithm as described in
http://people.csail.mit.edu/shanir/publications/disc2008_submission_98.pdf

Current implementation does not support resizing which makes it pretty
unsuitable and dangerous for arbitrary dynamic data. However, it should
be useful if keys do not change and are tested with particular size of
the container. This implementation also lacks features skipped in
previous version, such as providing iterators.

It is meant to be used as a basis for proper implementation, and
functions implemented are complete.